Remove cyclic import based on atom.py
Partially addresses https://github.com/jensengroup/propka-3.1/issues/49
